chore(venv): split out shellcmdgetter
